refactor(pin color and sort type structure for improved organization)
