+++ app/src/main/java/com/dudoji/android/pin/domain/PinColor.kt
@@ -1,3 +1,3 @@
-package com.dudoji.android.pin.color
+package com.dudoji.android.pin.domain
 
 data class PinColor(val id: Int, val name: String, val imageResId: Int)
